Updated Doctor service to include prescription functions since they only be used in that view. Also updated doctor scene to reflect changes to Physical_Test_Findings
